# openapi.yaml
openapi: 3.1.0
info:
  title: LLM Mock API Server
  version: 1.1.0
  description: |
    OpenAI-compatible mock server for testing streaming, reasoning, staged tool calling,
    and fault injection. Tool testing is driven through `/api/v1/chat/completions`.
servers:
  - url: http://localhost:8080
paths:
  /api/v1/chat/completions:
    post:
      summary: Create a chat completion
      description: |
        Supports `#CHAIN:` and `#CHAIN_STEPn:` directives in system messages.
        `#CHAIN_STEPn:` enables multi-round tool testing where the client executes
        real tools and sends `role=tool` messages back to continue the flow.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
      responses:
        '200':
          description: Chat completion or SSE stream
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
            text/event-stream:
              schema:
                type: string
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    ChatRequest:
      type: object
      required:
        - messages
      properties:
        messages:
          type: array
          items:
            $ref: '#/components/schemas/ChatMessage'
        model:
          type: string
        models:
          type: array
          items:
            type: string
        stream:
          type: boolean
        stream_options:
          $ref: '#/components/schemas/ChatStreamOptions'
        temperature:
          type: number
        top_p:
          type: number
        max_completion_tokens:
          type: integer
        max_tokens:
          type: integer
        reasoning:
          $ref: '#/components/schemas/ChatReasoningConfig'
        tools:
          type: array
          items:
            $ref: '#/components/schemas/ChatFunctionTool'
        tool_choice:
          description: Tool choice strategy passed through by the client.
        stop:
          description: Stop sequence configuration.
        seed:
          type: integer
        frequency_penalty:
          type: number
        presence_penalty:
          type: number
        logit_bias:
          type: object
          additionalProperties:
            type: number
        logprobs:
          type: boolean
        top_logprobs:
          type: integer
        user:
          type: string
        metadata:
          type: object
          additionalProperties:
            type: string
        parallel_tool_calls:
          type: boolean
        modalities:
          type: array
          items:
            type: string
        session_id:
          type: string
        service_tier:
          type: string
        debug:
          $ref: '#/components/schemas/ChatDebugOptions'
    ChatDebugOptions:
      type: object
      properties:
        deterministic:
          type: boolean
          description: Make generated ids deterministic when used with `seed`.
        chain_trace:
          type: boolean
          description: Emit selected chain-step trace information via response headers.
    ChatReasoningConfig:
      type: object
      properties:
        enabled:
          type: boolean
        effort:
          type: string
        max_tokens:
          type: integer
        exclude:
          type: boolean
        summary:
          description: Provider-specific reasoning summary configuration.
    ChatStreamOptions:
      type: object
      properties:
        include_usage:
          type: boolean
    ChatMessage:
      type: object
      properties:
        role:
          type: string
          enum: [system, user, assistant, tool]
        content:
          description: String or provider-specific structured content.
        name:
          type: string
        tool_calls:
          type: array
          items:
            $ref: '#/components/schemas/ChatToolCall'
        tool_call_id:
          type: string
        reasoning:
          type: string
        refusal:
          type: string
    ChatFunctionTool:
      type: object
      required:
        - type
        - function
      properties:
        type:
          type: string
          enum: [function]
        function:
          $ref: '#/components/schemas/FunctionDefinition'
    FunctionDefinition:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        description:
          type: string
        parameters:
          description: JSON Schema for function arguments.
        strict:
          type: boolean
    ChatResponse:
      type: object
      properties:
        id:
          type: string
        object:
          type: string
        created:
          type: integer
        model:
          type: string
        choices:
          type: array
          items:
            $ref: '#/components/schemas/ChatChoice'
        usage:
          $ref: '#/components/schemas/ChatUsage'
        system_fingerprint:
          type: string
        service_tier:
          type: string
    ChatChoice:
      type: object
      properties:
        index:
          type: integer
        message:
          $ref: '#/components/schemas/ChatMessage'
        finish_reason:
          type: string
          enum: [tool_calls, stop, length, content_filter, error]
    ChatToolCall:
      type: object
      required:
        - id
        - type
        - function
      properties:
        id:
          type: string
        type:
          type: string
          enum: [function]
        function:
          $ref: '#/components/schemas/FunctionCall'
    FunctionCall:
      type: object
      required:
        - name
        - arguments
      properties:
        name:
          type: string
        arguments:
          type: string
    ChatUsage:
      type: object
      properties:
        prompt_tokens:
          type: integer
        completion_tokens:
          type: integer
        total_tokens:
          type: integer
        prompt_tokens_details:
          $ref: '#/components/schemas/PromptTokensDetails'
        completion_tokens_details:
          $ref: '#/components/schemas/CompletionTokensDetails'
    PromptTokensDetails:
      type: object
      properties:
        cached_tokens:
          type: integer
        cache_write_tokens:
          type: integer
        audio_tokens:
          type: integer
        video_tokens:
          type: integer
    CompletionTokensDetails:
      type: object
      properties:
        reasoning_tokens:
          type: integer
        audio_tokens:
          type: integer
        accepted_prediction_tokens:
          type: integer
        rejected_prediction_tokens:
          type: integer
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            message:
              type: string
            type:
              type: string
            code:
              type: string